Keep MoonLive scripts on the filesystem, not in every module
A scripted module carried its script as a fixed 1 KB array, plus a second copy to notice
edits — resident whether or not a script was loaded, so six modules held ~16 KB of a
classic ESP32's 320 KB for text that was mostly empty. The script now lives in a file; the
module holds its name, reads it into a right-sized buffer to compile, and frees it. Scripts
are bounded by the filesystem instead of by an array nobody can grow.

Performance: desktop 132 us/tick (7575 fps), esp32 2151 us/tick (464 fps).

Light domain
- A `script` control (~32 B) replaces the `source` textarea in all three bindings. The UI
  loads, edits and saves the file through the /api/file endpoints that already existed, so
  this needed no new backend surface. A fresh module reports "no script — set the script
  name" and renders nothing, rather than every new module compiling the same default.
- The rebuild check is a 4-byte FNV-1a of the script text, not a second copy of it. It only
  ever answered "did this change".
- Per-binding control-name pools are gone: the engine owns the names it publishes now, so
  three private copies of the same fact went with them.
- /moonlive/ is created on demand — the write endpoint does not make parent directories, so
  a first save on a fresh device failed with nowhere obvious to look.

Core
- The engine copies declared control NAMES out of the source before returning. They pointed
  into the source text, which the caller is now free to release the moment compile() ends —
  and does. A control briefly appeared named "\x05" before this was found.
- IrProgram's op array is heap-allocated and sized from a token count, RAII-owned (destructor
  frees, copy deleted). It was a ~2 KB stack member on a 12 KB main task, the same cost for a
  one-statement script as a full one — so growing it would have traded a compile limit for a
  stack overflow. SEVEN sequential statements used to fail; forty compile. kMaxIrOps 64 →
  4096 is now a sanity bound, not the working limit.
- Widening that count to uint16_t left four uint8_t loop counters iterating over it — three
  lowerers and IrProgram::hasInline — which wrapped at 256 ops and spun forever. On a device
  that is a watchdog reset from a script that merely got long. Bisected (60 statements fine,
  80 hung); the regression test HANGS when the fix is reverted, which is how it was checked.
- ParlioLedDriver asks the platform for its 65535-byte transfer cap rather than naming the
  number in the light domain, and an over-capacity frame reports the ceiling in lights per
  pin on both the reinit and tick paths — the KB figure was the one a user could not act on.

Tests
- A shared fixture writes each script to a file, so tests exercise the path that ships. It is
  thread-local: the concurrency test compiles from two threads, and a shared name buffer had
  them compiling each other's script.
- Tests that relied on a built-in default script now name one. There is no default any more.

Docs/CI
- MIGRATING: `source` is gone, so a persisted script is an unknown key and ignored — the entry
  says where to find the text (/.config/Layouts.json as "N.source") and how to restore it.
- The three module specs, and the plan's step 1 marked done with what actually shipped.

Verified on the desktop: a 16x12 scripted grid layout with a scripted lines effect, both
compiled from files written over the API, surviving a restart and reloading from persistence.
Not yet run on hardware — the boards were unreachable; that is next.

Flash: esp32 1762368, esp32s3-n16r8 1752992, esp32s31 2025600, esp32p4-eth 1603952,
desktop 1138184. Tests: 1326 cases.

+++ src/platform/esp32/moonlive_lower_riscv.cpp
@@ -46,7 +46,9 @@ size_t lowerToBytes(const IrProgram& ir, uint8_t* out, size_t cap) {
         return labels[id];
     };
 
-    for (uint8_t i = 0; i < ir.count; i++) {
+    // uint16_t, matching IrProgram::count: the op array is sized to the script now, so a
+    // uint8_t counter wrapped at 256 ops and looped forever instead of emitting.
+    for (uint16_t i = 0; i < ir.count; i++) {
         const IrInst& op = ir.ops[i];
         switch (op.op) {
             case IrOp::Const:  a.movImm(reg(op.dst), op.imm); break;
